restructure server/client code
