+++ src/client/components/Welcome.astro
@@ -1,6 +1,6 @@
 ---
-import astroLogo from "../assets/astro.svg";
-import background from "../assets/background.svg";
+import astroLogo from "~/client/assets/astro.svg";
+import background from "~/client/assets/background.svg";
 ---
 
 <div id="container">
