+++ src/client/layouts/BaseLayout.astro
@@ -1,5 +1,5 @@
 ---
-import "~/styles/global.css";
+import "~/client/styles/global.css";
 
 const { title } = Astro.props;
 const pageTitle = `${title} | rsvpme`;
