+++ src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import AppLayout from "~/layouts/AppLayout.astro";
+import AppLayout from "~/client/layouts/AppLayout.astro";
 
 const user = await Astro.locals.currentUser();
 ---
