+++ src/pages/sign-up.astro
@@ -1,6 +1,6 @@
 ---
 import { SignUp as ClerkSignUp, SignedOut } from "@clerk/astro/components";
-import AppLayout from "~/layouts/AppLayout.astro";
+import AppLayout from "~/client/layouts/AppLayout.astro";
 
 const { userId } = Astro.locals.auth();
 if (userId) {
